[refactor] type: drive mypy --strict to zero across workspace
Annotate src/mcapp (production + test suites) and ble_service/src to strict
cleanliness (267 -> 0 errors). Fixes read from the code, no behavior change
except on already-unreachable paths.

Genuine fixes surfaced by the type checker:
- meteo: widen _calculate_cloud_coverage_description(cloud_percent) int->float
  (body is round(x/12.5); enables exact okta banker's-rounding boundary tests)
- classifier/tests: replace stale dead mcapp.storage.Storage reference
- ble_adapter: guard two latent None-deref paths; close a _load_ble_state
  Any-leak in ble main
- contract_parity_tests: fix a latent variable-shadowing type bug
- udp_handler: correct a stale/incorrect method-assign type: ignore
- storage/_base: declare missing cross-mixin stubs (_build_message_dict,
  get_blocked_texts); widen _execute_many to Sequence[tuple]

All 14 residual `# type: ignore` are scoped `[code]` with a reason.
scripts/run_startup_tests.py exits 0; ruff check/format clean.

+++ src/mcapp/meteo.py
@@ -657,7 +657,7 @@ class WeatherService:
 
     def _calculate_cloud_coverage_description(
         self,
-        cloud_percent: int | None,
+        cloud_percent: float | None,
         timestamp_str: str | None = None,
     ) -> str:
         """Berechne Wolkenbedeckung in Achteln (/8) und Beschreibung"""
